kubernetes charts examples and jenkins
We are maintaining our own Jenkins server on Kubernetes in jenkins-git

It is derived from kubernetes/charts stable/jenkins chart, but kept
current as we upgrade our jenkins master chart via the helm upgrade.

An example from examples might be used to handle future Jenkins server
that can be rotated in and out of service during rolling deployment.

The nfs-pv-provider is supposed to facilitate that, since ReadWriteMany
won't need to wait for incumbent server to resign before master-elect
can attach the PV disk to finish Pending and go into ContainerCreating

It might work or it might be absolutely terrible idea...